Add more robust triangle function .
The whole algorithm should reconsidered, should include a lot of extreme situation like
point on the vertex, on the boundary edge should add more cache to keep the topology un-contaminate
topology consistence keeping is very difficult task. but when this is done the algorithm may faster and robust than other geometry methods!